+++ target-sparc/op_helper.c
@@ -3612,7 +3612,7 @@ static void cpu_restore_state2(void *retaddr)
 
     if (retaddr) {
         /* now we have a real cpu fault */
-        pc = (unsigned long)retaddr;
+        pc = (uintptr_t)retaddr;
         tb = tb_find_pc(pc);
         if (tb) {
             /* the PC is inside the translated code. It means that we have
