+++ targphys.h
@@ -6,7 +6,7 @@
 #ifdef TARGET_PHYS_ADDR_BITS
 /* target_phys_addr_t is the type of a physical address (its size can
    be different from 'target_ulong'). We have sizeof(target_phys_addr)
-   = max(sizeof(unsigned long),
+   = max(sizeof(uintptr_t),
    sizeof(size_of_target_physical_address)) because we must pass a
    host pointer to memory operations in some cases */
